+++ src/test/suite/index.ts
@@ -1,0 +1,33 @@
+import * as path from 'path';
+import Mocha from 'mocha';
+import { glob } from 'glob';
+
+export async function run(): Promise<void> {
+  // Create the mocha test runner
+  const mocha = new Mocha({
+    ui: 'bdd',
+    color: true,
+    timeout: 10_000,
+  });
+
+  const testsRoot = path.resolve(__dirname, '.');
+
+  const files = await glob('**/*.test.js', { cwd: testsRoot });
+
+  // Add all test files to the mocha instance
+  files.forEach((f) => mocha.addFile(path.resolve(testsRoot, f)));
+
+  return new Promise((resolve, reject) => {
+    try {
+      mocha.run((failures) => {
+        if (failures > 0) {
+          reject(new Error(`${failures} test(s) failed.`));
+        } else {
+          resolve();
+        }
+      });
+    } catch (err) {
+      reject(err);
+    }
+  });
+}